Move existing "fun" things into the `typeclass` package
New package helps to differentiate "typeclass definitions" from other
related definitions (like instances of the typeclasses).
