+++ json/parser/src/TermParser.scala
@@ -1,7 +1,7 @@
 package io.github.maxkar
 package json.parser
 
-import fun.Monad
+import fun.typeclass.Monad
 
 /**
  * Parser for simple term value (true, false, boolean).
