+++ json/parser/test/QualitativeParser.scala
@@ -1,7 +1,7 @@
 package io.github.maxkar
 package json.parser
 
-import fun.Monad
+import fun.typeclass.Monad
 
 /**
  * Qualitative parser definitions - only indicates if parsing of a given string was successfull or not.
