+++ json/parser2/src/Arrays.scala
@@ -1,7 +1,7 @@
 package io.github.maxkar
 package json.parser
 
-import fun.Monad
+import fun.typeclass.Monad
 import scala.collection.mutable.ArrayBuffer
