+++ json/parser2/src/Errors.scala
@@ -1,7 +1,7 @@
 package io.github.maxkar
 package json.parser
 
-import fun.Monad
+import fun.typeclass.Monad
 
 /**
  * Error-related utilities.
