+++ json/parser2/src/Literals.scala
@@ -1,8 +1,8 @@
 package io.github.maxkar
 package json.parser
 
-import fun.Monad
-import fun.Applicative
+import fun.typeclass.Monad
+import fun.typeclass.Applicative
 
 /**
  * Literal readers.
