+++ json/simple/factory/src/Factory.scala
@@ -9,8 +9,7 @@ import json.parser.factory.NumberFactory
 import json.parser.factory.ArrayFactory
 import json.parser.factory.ObjectFactory
 
-import fun.Monad
-import scala.compiletime.ops.string
+import fun.typeclass.Monad
 
 /**
  * Json object/value factory compatible with the standard parser API.
