+++ json/simple/query/test/MonadicConversionTest.scala
@@ -8,8 +8,8 @@ import json.simple.Json
 import json.simple.query.given
 import defaultConversions.given
 
-import fun.Monad
-import fun.Collect
+import fun.typeclass.Monad
+import fun.typeclass.Collect
 
 import scala.language.implicitConversions
